legacy/lava/__init__.py: Do not overwrite existing job file
If we just overwrite existing job file due name collision we are
creating hard to troubleshoot cases, when some LAVA tests are missing.
It is better to fail, if file already exist, so it will be visible.

+++ kernelci/legacy/lava/__init__.py
@@ -10,6 +10,7 @@
 from jinja2 import Environment, FileSystemLoader
 import json
 import os
+import sys
 
 from kernelci.runtime import add_kci_raise
 
@@ -77,6 +78,11 @@ class LavaRuntime:
     def save_file(self, job, output_path, params):
         file_name = self.job_file_name(params)
         output_file = os.path.join(output_path, file_name)
+        if os.path.isfile(output_file):
+            # print error message to stderr
+            print("ERROR: file already exists: %s" % output_file,
+                  file=sys.stderr)
+            return None
         with open(output_file, 'w') as output:
             output.write(job)
         return output_file
